% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnas_data.R
\name{pnas_data}
\alias{pnas_data}
\title{Construct economic data used in PNAS paper}
\usage{
pnas_data(prairie_cost_per_acre = 95)
}
\arguments{
\item{prairie_cost_per_acre}{A scalar indicating how much the prairie per
acre in dollars.}
}
\value{
A \code{data.frame} containing the columns PI, source, watershed,
year, response, and value.
}
\description{
This function will calculate the revenue for a watershed-year combination
for watersheds that have a certain amount of prairie.
}
\examples{
d <- pnas_data()
summary(d)

}
