% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revenue}
\alias{revenue}
\title{Revenue data}
\format{A \code{data.frame} with 3 variables: watershed, year, and revenue}
\usage{
revenue
}
\description{
Revenue data for crop portion of watersheds, i.e. it does not account for the
fact that some of the watersheds have a portion that is prairie. For details
on how these revenue values were obtained, see the xlsx file in the data-raw
folder.
}
\seealso{
\code{\link{pnas_data}}
}
\keyword{datasets}
